fix: unify the zero-ppid guard, split macos_tests.rs, pin the sysctl-fallback test's precondition
