+++ src/identity.rs
@@ -150,9 +150,10 @@ impl ProcessId {
 #[cfg(windows)]
 pub(crate) use backend::{close as windows_close, open_classified as windows_open_classified, Opened};
 
-/// `containment::enumerate::macos`'s sysctl fallback for a pid's parent pid — see
-/// `identity::macos::ppid_of`'s doc for why it lives here instead of a second copy of
-/// `kinfo_proc`'s layout.
+/// `containment::enumerate::macos`'s parent-pid resolver, reused whole (primary
+/// `proc_pidinfo` read, sysctl fallback, and the shared zero-ppid guard) — see
+/// `identity::macos::ppid_of`'s doc for why it lives here instead of a second copy of either
+/// the guard or `kinfo_proc`'s layout.
 #[cfg(target_os = "macos")]
 pub(crate) use backend::ppid_of as macos_ppid_of;
